pyproject.toml
[project]
name = "shu-agentic-rag-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.1",
    "beautifulsoup4>=4.14.3",
    "chroma>=0.2.0",
    "chromadb>=1.5.5",
    "datasets>=4.8.4",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "jq>=1.11.0",
    "langchain>=1.2.12",
    "langchain-aws>=1.4.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.2",
    "lxml>=6.0.2",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.9",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "requests>=2.32.5",
    "selenium>=4.41.0",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "webdriver-manager>=4.0.2",
]
